+++ pos_multi_session/models/pos_multi_session_models.py
@@ -24,10 +24,6 @@ class PosMultiSession(models.Model):
     _name = 'pos.multi_session'
 
     name = fields.Char('Name')
-    remove_unpaid_orders = fields.Boolean('Remove unpaid orders', default=True, help="If the parameter is True, all unpaid orders will be "
-                                                                                     "removed from server and order number will reset to 1 again, "
-                                                                                     "once all synced sessions are closed. Otherwise, all such orders "
-                                                                                     "will stay on server regardless of sessions closing.")
     pos_ids = fields.One2many('pos.config', 'multi_session_id', 'POSes')
     order_ids = fields.One2many('pos.multi_session.order', 'multi_session_id', 'Orders')
     order_ID = fields.Integer(string="Order number", default=0, help="Current Order Number shared across all POS in Multi Session")
@@ -141,7 +137,7 @@ class PosMultiSessionOrder(models.Model):
 
     order = fields.Text('Order JSON format')
     order_uid = fields.Char(index=True)
-    state = fields.Selection([('draft', 'Draft'), ('deleted', 'Deleted')], default='draft', index=True)
+    state = fields.Selection([('draft', 'Draft'), ('deleted', 'Deleted'), ('unpaid', 'Unpaid and removed')], default='draft', index=True)
     revision_ID = fields.Integer(default=1, string="Revision", help="Number of updates received from clients")
     multi_session_id = fields.Many2one('pos.multi_session', 'Multi session', index=True)
 
@@ -152,13 +148,10 @@ class PosSession(models.Model):
     @api.multi
     def action_pos_session_close(self):
         res = super(PosSession, self).action_pos_session_close()
-        remove_unpaid_orders = self.config_id.multi_session_id.remove_unpaid_orders
-        if remove_unpaid_orders:
-            self.config_id.write({'multi_session_message_ID': 1})
-            active_sessions = self.env['pos.session'].search([('state', '!=', 'closed'), ('config_id.multi_session_id', '=', self.config_id.multi_session_id.id)])
-            if len(active_sessions) == 0:
-                self.config_id.multi_session_id.sudo().write({'order_ID': 0})
-                orders = self.config_id.multi_session_id.order_ids.search([('state', '!=', 'deleted')])
-                for order in orders:
-                    order.state = 'deleted'
+        self.config_id.write({'multi_session_message_ID': 1})
+        active_sessions = self.env['pos.session'].search([('state', '!=', 'closed'), ('config_id.multi_session_id', '=', self.config_id.multi_session_id.id)])
+        if len(active_sessions) == 0:
+            self.config_id.multi_session_id.sudo().write({'order_ID': 0})
+            orders = self.config_id.multi_session_id.order_ids.filtered(lambda x: x.state == "draft")
+            orders.write({'state': 'unpaid'})
         return res