+++ pos_multi_session/static/src/js/pos_multi_session.js
@@ -51,7 +51,7 @@ odoo.define('pos_multi_session', function(require){
             this.get('orders').bind('remove', function(order, collection, options){
                 if (!self.multi_session.client_online) {
                     if (order.order_on_server ) {
-                        var warning_message = _t("No connection to the server. You can only create new orders. It is forbidden to modify existing orders.");
+                        var warning_message = _t("No connection to the server. You can create new orders only. It is forbidden to modify existing orders.");
                         self.multi_session.warning(warning_message);
                         return false;
                     }
@@ -443,7 +443,6 @@ odoo.define('pos_multi_session', function(require){
             this.order_ID = null;
             this.update_queue = $.when();
             this.func_queue = [];
-
             this.pos.longpolling_connection.on("change:poll_connection", function(status){
                 if (status) {
                     if (self.offline_sync_all_timer) {
@@ -452,9 +451,11 @@ odoo.define('pos_multi_session', function(require){
                     }
                     self.request_sync_all();
                 } else {
-                    var warning_message = _t("No connection to the server. You can only create new orders. It is forbidden to modify existing orders.");
-                    self.warning(warning_message);
-                    self.start_offline_sync_timer();
+                    if (!self.offline_sync_all_timer) {
+                        var warning_message = _t("No connection to the server. You can create new orders only. It is forbidden to modify existing orders.");
+                        self.warning(warning_message);
+                        self.start_offline_sync_timer();
+                    }
                 }
             });
         },
@@ -504,7 +505,12 @@ odoo.define('pos_multi_session', function(require){
                 if(error.message === 'XmlHttpRequestError ') {
                     self.client_online = false;
                     e.preventDefault();
-                    self.pos.longpolling_connection.set_status(false);
+                    self.pos.longpolling_connection.network_is_off();
+                    if (!self.offline_sync_all_timer) {
+                        var warning_message = _t("No connection to the server. You can create new orders only. It is forbidden to modify existing orders.");
+                        self.warning(warning_message);
+                        self.start_offline_sync_timer();
+                    }
                 } else {
                     self.request_sync_all();
                 }
@@ -512,8 +518,6 @@ odoo.define('pos_multi_session', function(require){
                 if (self.pos.debug){
                     console.log('MS', self.pos.config.name, 'response #'+current_send_number+':', JSON.stringify(res));
                 }
-                self.pos.longpolling_connection.set_status(true);
-
                 var server_orders_uid = [];
                 self.client_online = true;
 
@@ -529,8 +533,14 @@ odoo.define('pos_multi_session', function(require){
                     });
                     self.pos.pos_session.order_ID = res.order_ID;
 
-                    // self.pos.pos_session.sequence_number = res.order_ID;
+                    if (res.order_ID != 0) {
+                        self.pos.pos_session.sequence_number = res.order_ID;
+                    }
                     self.destroy_removed_orders(server_orders_uid);
+                }
+                if (self.offline_sync_all_timer) {
+                    clearInterval(self.offline_sync_all_timer);
+                    self.offline_sync_all_timer = false;
                 }
             });
         },